requirements.txt
# Cognitive-Aim Experiment B - Minimal Dependencies
# Python >= 3.9 recommended

# Core ML frameworks
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.21.0
Pillow>=8.3.0

# Vision transformers and pretrained models
timm>=0.9.0
transformers>=4.20.0

# Data handling
opencv-python>=4.5.0
scipy>=1.7.0
matplotlib>=3.5.0

# Configuration and utilities
PyYAML>=6.0
tqdm>=4.62.0

# Optional: For CUDA acceleration
# Install CUDA-compatible versions:
# pip install torch torchvision --index-url https://download.pytorch.org/whl/cu118

# Optional: For advanced metrics
scikit-image>=0.19.0